Revert "rename account-setting module, add service layer"
This reverts commit 11b5ab7439b1b9b745f878144b924b5116e6a872.
