+++ webClient/src/app/app.module.ts
@@ -13,7 +13,6 @@ import { FooterComponent } from './components/footer/footer.component';
 import { FormsModule } from '@angular/forms';
 import { DynamicFormComponent } from './share/dynamic-form/dynamic-form.component';
 import { DynamicFormFieldComponent } from './share/dynamic-form-field/dynamic-form-field.component';
-import { WebService } from './service/web.service';
 
 
 
@@ -31,7 +30,7 @@ import { WebService } from './service/web.service';
     FormsModule,
     Router
   ],
-  providers: [WebService],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
