+++ webClient/src/app/app.router.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { AccountSettingRegistrationComponent } from './components/account-setting-registration/account-setting-registration.component';
+import { AccountSettingComponent } from './components/account-setting/account-setting.component';
 import { SearchComponent } from './components/search/search.component';
 import { AccountRecoverComponent } from './components/account-recover/account-recover.component';
 import { ProfileComponent } from './components/profile/profile.component';
@@ -8,20 +8,20 @@ import { ProfileComponent } from './components/profile/profile.component';
 
 const routes = [
     {path : '', component : SearchComponent},
-    {path : 'register', component : AccountSettingRegistrationComponent},
-    {path : 'account-setting/:id', component : AccountSettingRegistrationComponent},
+    {path : 'register', component : AccountSettingComponent},
+    {path : 'account-setting/:id', component : AccountSettingComponent},
     {path : 'account-recover', component : AccountRecoverComponent},
     {path : 'profile', component : ProfileComponent}
 ];
 
 @NgModule({
-    imports: [
+    imports:[
         RouterModule.forRoot(routes)
     ],
-    exports: [
+    exports:[
         RouterModule
     ]
 })
 
-export class Router {}
-export const routingComponets = [AccountSettingRegistrationComponent, SearchComponent, AccountRecoverComponent, ProfileComponent]
+export class Router{}
+export const routingComponets = [AccountSettingComponent, SearchComponent, AccountRecoverComponent, ProfileComponent]
