+++ webClient/src/app/components/account-setting/account-setting.component.html
@@ -1,0 +1,6 @@
+<div class="accountSetting">
+    <!-- <span ngif="id == null">Account id : {{id | async}}</span> -->
+    <app-dynamic-form [dataObj]="mock_data" [formatObj]="formatMap"></app-dynamic-form>
+    <button (click)='onSave()'>Save</button>
+    <div class="clear-fix"></div>      
+</div>
