+++ webClient/src/environments/environment.ts
@@ -4,7 +4,5 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 export const environment = {
-  name : 'local',
-  production : false,
-  API_URL : 'http://localhost:8080'
+  production: false
 };
