+++ webClient/src/styles.css
@@ -7,13 +7,9 @@ body {
     background-color: #fafafa;
 }
 
-.title {
-    font-size: 1.6rem;
-    color: #666;
-    padding: 10px 0;
+.clear-both {
+    clear : both;
 }
-
-
 
 button {
     padding: 10px 15px;
@@ -37,8 +33,4 @@ button:hover {
 button:focus,
 input:focus {
     outline:0;
-}
-
-.clear-both {
-    clear : both;
 }
